commands/task-extractor/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-pubsub==2.*
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
tzdata>=2024.1
cloudevents>=1.9.0
